fix(tags): Set consistent titles for tags with description pages
Right now for some tag `my-tag` a tag page with title `Tag: my-tag` is
generated.
However if you create a page `tags/my-tag.md` to add a description the
page will have the title `my-tag` by default.

Set the tag page title to have the `Tag: ` prefix if the tag page does
not have a custom title for consistency between tags with and without
descriptions.